requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
requests
openai
python-dotenv
psycopg[binary]>=3.2.2,<4.0
stripe>=10.0.0
httpx

